Modify die interface.
A die now has a `roll` method in addition to the die itself being able to be evaluated using the function call operator.  A die now also carries a `sides` property.

With these changes, it is no longer necessary to return a compound object when rolling a die.  Rolling a die can now simply return the result of the roll.