Change dice expression result to be a hierarchy of objects rather than a single string.
A dice expression result still evaluates to a single number via the `value()` method.  However, we must keep track of the result at each node of the originating expression in order to properly format the result.  For example, die rolls will be formatted to include both the roll (e.g. 4) and the die that was rolled (e.g. d6).