Modify dice expression parser tests to compare result with expected expression rather than expected result value.
Note that this required a fundamental change to the type of a dice expression.  Dice expressions previously were functions with additional members.  However, it isn't possible to compare functions for equality due to the use of closures.  Therefore, the ability to invoke a dice expression directly (i.e. using `()`) has been removed; the use of `evaluate()` is now the only mechanism for evaluating the expression.