Associate custom dice bag with each dice expression parser instance instead of using a global instance for all parsers.
Ideally, I would like to be able to pass such parameters directly to the Jison-generated `Parser` class ctor, but this doesn't appear to be possible at this time (see Jison issue [241](https://github.com/zaach/jison/issues/241)).  Instead, I export a factory method fromm the parser module that adds the dice bag to the `yy` property of the `Parser`.  This was necessary because `yy` appears to be the only link to the outside world available for use in the parser grammar rules.