Force dice roll literals to have an explicit count.
This change is necessary because a forthcoming change will require that we be able to distinguish between dice roll literals (e.g. `3d6) and die literals (e.g. 'd6').  Die literals are required so one can pass a die for rolling to functions (e.g. `roll()`).

I can't think of a better syntax for a die literal at this point that is intuitive.  Thus, we sacrifice some convenience for a richer parser grammar (e.g. having to enter `1d6` instead of `d6` to roll a single six-sided die).