Ensure response signature algorithm is cross-platform.
Per MDN docs for `JSON.stringify`:

"Properties of non-array objects are not guaranteed to be stringified in any particular order. Do not rely on ordering of properties within the same object within the stringification."

The root cause of this problem is due to the lack of a standardized canonicalized form for JSON.

Therefore, we must choose a canonicalization algorithm that is consistent across all platforms.  We choose to base64-encode the JSON text to obtain a canonicalized form for which the signature can be calculated consistently.  Base64 encoding of the JSON object to be signed appears to be the mechanism forthcoming JSON signature standards are proposing.

The downside of this approach is that the JSON text is no longer human readable with respect to the most interesting content.  Thus, debugging will require an additional decoding step.