Temporary fix for broken acceptance tests.
Cucumber merges all Given/When/Then pattern rules under the directory from which it is run into a single instance.  If any rules have a matching pattern, the last one registered wins.  This can cause the wrong rule to be invoked resulting in a test failure.

This temporary fix simply runs each independent set of acceptance tests separately so there's no chance that their rules will get merged together.  A solution that doesn't smell so bad will be pursued in the future.