Always use server's public key to verify signatures.
The `node-jws` module does not yet implement X.509 certificate chain validation.  Thus, it is not safe to use the public key embedded in the JWS protected header as it may be from an untrusted source.  Until certificate chain validation is implemented, we must explicitly use the server's public key to verify a signature.

An implication of this is that a load-balanced service would have to use the same public/private key pairs across all service instances.