Use HTTP response status codes to indicate success/failure.
Per Chapter 11 in RESTful Web APIs, the success or failure of a response should not be conveyed in the entity-body, but rather through the status code.  This change removes the "success" and "failure" objects in the response body and instead returns 200 (OK) for success and 404 (Bad Request) for failure.

A future change will add support for returning 500 (Internal Server Error) when a true runtime error occurs as opposed to an error caused by the caller's request being malformed.