Require RNG specification to be signed.
The RNG specification is really meant to only be used by test code to ensure all results are deterministic.  However, it could be hijacked by a rogue GM to ensure the players rolls evaluate to a specific value.  With this change, if an RNG specification is present in an evaluate expression request, it must be signed by a trusted authority.  Currently, only the dice server itself is a trusted authority.