Use `random-js` module for random number generation.
We are primarily using this module for two reasons:

1) We want to be able to use an RNG that supports seeding.  A future change is going to include an RNG seed in a ticket to ensure redeeming a ticket always generates the same expression result no matter how many times it is redeemed (and relieves the service from having to persist the ticket IDs it has already redeemed).

2) We want to avoid bias in die rolls as much as possible.  Our previous algorithm was subject to bias as described at <https://github.com/ckknight/random-js>.

Using this module required a significant change in how we modeled RNGs.  Previously, an RNG was a zero-argument function that returned a value in range [0,1).  Now, an RNG is a one-argument function that returns a value in the range [1,`sides`], where `sides` is the single RNG argument.  This is due to how `random-js` generates a random number from both an engine and a distribution.  We cannot simply stub the engine to generate a specific random number using the `integer` distribution because the distribution algorithm does not use a simple mapping function to go from the engine value to a value in the range [1,`sides`].  Therefore, the entire engine/distribution combination must be abstracted behind the "RandomNumberGenerator" interface.