No longer track redeemed tickets.
Now that a ticket includes an RNG seed, it is guaranteed that redeeming a ticket is idempotent.  Therefore, it is no longer necessary to track which tickets have been redeemed and to report an error if a player attempts to redeem a ticket more than once.  Because of idempotence, it is not possible for a player to keep trying to redeem a ticket until they get a roll they like.