Fix unexpected redirection error in Travis build
The version of bash on Travis (Ubuntu 12.04.5 LTS) apparently does not
support redirecting the output of command group executed in the context
of the current shell (i.e. using curly braces):

  /bin/sh: 10: Syntax error: redirection unexpected

Switched to parentheses to cause a subshell to be created, as that may
be a requirement in the version of bash on Travis.