Fix intermittently-failing acceptance tests
As of 193492f, the HTML fragments for each module have been extracted to
an external file.  These fragments are now loaded asynchronously.  Due
to this asynchronous loading, there is now a race condition when
Selenium attempts to locate elements within those fragments.

We fix this by explicitly waiting for all fragments to be loaded before
informing Cucumber that the home page is loaded.  We probably need a
better solution in the future because this solution will not scale as
more feature fragments are added.