Fix race conditions in client acceptance tests
After changing the build system to Gulp, several client acceptance tests
were observed to intermittently fail.  It is believed the Gulp Cucumber
plugin caused the timing of the tests to change (presumably they are now
running faster because the existing Node process is being forked rather
than spinning up an entirely new Node process via the shell), which
exacerbated the existing race conditions.

The typical race condition involved the Cucumber steps (both When and
Then steps) racing ahead of the expected state of the UI.  For example,
after evaluating two expressions, the test would try to click the Remove
button of the first row in the results table.  However, when the test
loses the race, the second row would have not yet been added.  Thus, the
wrong row was being removed.

This fix changes all assertions to effectively wait for a short period
of time for the assertion to be satisfied.  If the assertion is not
satisfied during that window, it will fail as before as if it were run
synchronously.