Improve race condition fix
This change partially reverts the previous fix for race conditions in
the client acceptance tests so that the tests once again read more like
a synchronous flow of events.  Now the synchronous promise chain is
passed to a single method that continuously evaluates it until it is
resolved.  If it is not resolved within the timeout, it will be
evaluated one final time, at which point it may be rejected.  If
rejected, the result will be the same as if the promise chain were
evaluated without waiting (i.e. any failed expectation error raised by
Chai will be reported as one would expect).