Prevent ESLint from ignoring files in the ".build" directory
This is a result of a previous change where we renamed the build output
directory from "build" to ".build".  By default, ESLint adds an ignore
rule for ".*".  Thus, any file within the ".build" directory is ignored.

The fix is to specify "dotfiles: true" in the ESLint configuration.
However, this does not appear to be an officially documented option for
the CLIEngine class (and it is not even available via the CLI).  Rather,
it is documented as part of the IgnoredPaths class, and we just happen
to be able to specify it as part of the CLIEngine options so that it
gets passed through to IgnoredPaths.  This seems a bit brittle and we
should consider renaming the build output directory once again to
something that does not use a leading dot.

In addition, I enabled a warning to be emitted any time ESLint ignores a
file.