Fix race condition in client acceptance tests
The past several fixes we've had to implement indicate, in certain
scenarios, we need to wait for the expression evaluation to complete
before evaluating the next expression.  However, this technical
requirement should not be visible in the specifications.  A future
change will attempt to hide these non-business-related steps within the
step definition for evaluating an expression.