Improve past fixes for race conditions in client acceptance tests
Improve the specifications to hide the fact that we need to wait for a
new row to appear in the history table.  Now we simply state that we are
expecting the expression evaluation to be successful.  The step
definition takes care of the technical detail of waiting for a new row
to appear if the claim of a successful evaluation is made.