Pass globals as module parameters instead of referencing them directly
I find this variant of the Revealing Module pattern to be a bit cleaner
as it makes it explicit which external variables are being referenced
within a module.

I modified the ESLint configuration to disallow all globals.  The only
drawback is this requires explicitly disabling the "no-undef" rule in
every module on the line where the IIFE arguments are specified.  Not
sure if there's a better way to do this.