Fix race conditions in client acceptance tests
The newly-revealed race conditions are a direct result of removing the
previous wait in HomePage.prototype.open for the individual HTML
fragments to load (because we no longer load them asynchronously).

However, the WebDriver.prototype.open method returns as soon as the
document is loaded.  In addition, the now-synchronous initialization
script runs as soon as the document is loaded.  The order in which these
two subscribers are informed of the document loaded event appears to
have changed after we inlined the CSS modules.  Thus, new race
conditions that assumed certain controls (contained within the HTML
fragments) were present were revealed.