Fix intermittently failing client acceptance test
The scenario "Evaluating empty expressions" in the "Evaluating dice
notation via web UI" feature was failing intermittently when run in
PhantomJS.  The root cause of the failure appears to be that the client
was submitting a blank ("") expression to the server, which caused the
error message to be displayed, thus failing the test.

It is assumed that PhantomJS is not properly honoring the "required"
attribute of the expression input field under some (as yet undetermined)
condition.  When this happens, the client submits the blank expression
to the server, and the server returns an error, which is then displayed.
This is not a race condition because, no matter how long the timeout is
increased to, the error message is always displayed and never removed.

The fix is to ensure the client never sends a falsey expression to the
server.  This serves as a second line of defense in the event the form
validation logic does not work (as happens with PhantomJS).  No
additional logic is performed to ensure the user sees a failure message
because this issue has never been observed in either Firefox or Chrome.
As it only affects PhantomJS, and PhantomJS is headless, attempting to
display some kind of feedback to the user would be pointless.