Attempt to fix intermittent PhantomJS acceptance test failures
The root cause of the problem seems to be that, when evaluating multiple
expressions, the response for the second request could be received
before the response to the first request.  This was possible because we
weren't waiting for the response to the first request before sending the
second request.  In essence, multiple requests could be in-flight at the
same time.

The code previously assumed that responses were returned in the order
they were sent.  In this change, we now wait for the response to a
request to be received before sending the next request.  This is done
via the newly-added support for request and correlation IDs.

We removed the previous functionality for waiting for a new row to
appear in the history table in certain tests before submitting the next
request.  To temporarily address the regressions caused by the removal
of this functionality, we once again added new steps that explicitly
wait for the number of rows in the history table to reach a particular
count.  We will once again attempt to remove these explicit steps in a
future change in order to hide them from the acceptance criteria.