Improve step definition that waits for results table rows
Replace the explicit step for waiting for a specific count of results
table rows with a new clause on the evaluation step that has the same
effect.  This helps to hide (but isn't perfect) this technical detail
required to avoid a race condition in the acceptance tests from the
specification writer.