Use separate variables for Heroku public/private keys
Otherwise the acceptance tests that run during the core build process
will use a different set of encryption keys (from the environment
variables) than are used by the production code (from the command line).

When running the acceptance tests against Heroku, we simply copy over
the Heroku-specific environment variables to the standard environment
variables only for the duration of the tests.