Include build system qualifier in version string
Because we've chosen to bump the production version immediately before
merging to the master branch, this can lead to confusion when viewing
the application in the staging environment.  Specifically because the
version displayed will be the previously-released version, which we will
have (presumably) since augmented.

The solution is to include specific information about the build itself.
When built from Travis, we include the branch name and the build number
(the branch name is not actually necessary because Travis assigns a
unique build number across all branches, but it makes the format
consistent with the non-Travis build, described next).  When built
locally, we include the branch name and the commit ref of the local Git
repo.