Add support for true percentile dice
Most users think of "d%" (i.e. percentile dice) to be rolled using two
10-sided dice, where one represents the tens value and the other
represents the ones value.  We previously mapped "1d%" to simply
"1d100".

In order to provide the option to roll true percentile dice, "d%" is now
equivalent to "2d10", with the additional logic of calculating a value
in the range [1,100] from the rolls of the two 10-sided dice.

A major difference from the prior implementation is that "d%" can be
entered directly; it does not need to be entered as "1d%" as was
required previously.  This is because there is no longer any aliasing
between a die literal and a dice roll literal now that we've changed the
definition of what "d%" means. Not to mention, I've never seen a game
use an expression like "2d%", so I'm not sure that should have ever been
a valid literal.