Use forwarded protocol if available when constructing endpoint URLs
Heroku provides an HTTPS front-end to our HTTP application.  When
constructing redeem and validate URLs, we should direct the user to the
endpoint using the same protocol they used when accessing Heroku.
Therefore, use the "X-Forwarded-Proto" header to define the endpoint URL
protocol if it's available.  Otherwise, fall back to the protocol
defined in the Express request.