Use gitignore to exclude files from processing
This significantly shortens many of our gulp.src() calls as we no longer
have to whitelist everything.

However, I have noticed a significant increase in the time to run the
affected tasks.  I believe this is because our initial glob includes
deep folders, such as node_modules, and every matching file then has to
pass through gulp-exclude-gitignore to be filtered.  So there's a lot of
CPU busy work going on now.

A possible solution in the future would be to add an additional API to
gulp-exclude-gitignore (or a totally new package because it would no
longer be Gulp-specific) that takes a glob pattern (either a String or
an Array) and appends negative globs corresponding to gitignore.
gulp-exclude-gitignore already basically does this, but it simply
passes on those strings to gulp-ignore; it provides no visibility to
this functionality via the API.